Add IT tests
