+++ core/src/test/java/io/nuun/kernel/core/pluginsit/dummy1/InterfaceWithCustom2Suffix.java
@@ -21,16 +21,8 @@ package io.nuun.kernel.core.pluginsit.dummy1;
 
 import javax.annotation.Nullable;
 
-
-
-/**
- * @author Epo Jemba
- *
- */
 @Nullable
-//@Ignore
 public interface InterfaceWithCustom2Suffix
 {
-    
     public String name();
 }
