+++ core/src/test/java/it/fixture/scan/ClassNotToScan1.java
@@ -1,0 +1,11 @@
+package it.fixture.scan;
+
+import io.nuun.kernel.api.annotations.Ignore;
+
+/**
+ * @author pierre.thirouin@ext.mpsa.com (Pierre Thirouin)
+ */
+@ToScan
+@Ignore
+public class ClassNotToScan1 {
+}
