+++ core/src/test/java/it/fixture/scan/ClassNotToScan2.java
@@ -1,0 +1,9 @@
+package it.fixture.scan;
+
+/**
+ * @author pierre.thirouin@ext.mpsa.com (Pierre Thirouin)
+ */
+@ToScan
+@CustomIgnore
+public class ClassNotToScan2 {
+}
