+++ core/src/test/java/it/fixture/scan/ClassToScan2.java
@@ -1,0 +1,8 @@
+package it.fixture.scan;
+
+/**
+ * @author pierre.thirouin@ext.mpsa.com (Pierre Thirouin)
+ */
+@ToScan
+public class ClassToScan2 {
+}
