+++ core/src/test/java/it/fixture/scan/ToScan.java
@@ -1,0 +1,14 @@
+package it.fixture.scan;
+
+import java.lang.annotation.*;
+
+/**
+ * Used for test fixtures. Indicates classes to scan.
+ *
+ * @author pierre.thirouin@ext.mpsa.com (Pierre Thirouin)
+ */
+@Documented
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.TYPE, ElementType.ANNOTATION_TYPE})
+public @interface ToScan {
+}
