+++ core/src/test/resources/META-INF/services/io.nuun.kernel.api.Plugin
@@ -1,1 +1,2 @@
-io.nuun.kernel.core.pluginsit.dummy1.DummyPlugin
+it.fixture.NoOpPlugin1
+it.fixture.NoOpPlugin2
